Fix: problem DELETE teams
fixed the problem that if a team had associated players it was not eliminated correctly and if it was associated to a match, in the latter case it is not eliminated since it is not possible to eliminate teams already assigned to soccer matches.

+++ backend/src/main/java/es/codeurjc/backend/RESTController/TeamRestController.java
@@ -1,11 +1,17 @@
 package es.codeurjc.backend.RESTController;
 
+import com.fasterxml.jackson.databind.deser.DataFormatReaders;
 import es.codeurjc.backend.DTOs.TeamDTO;
+import es.codeurjc.backend.model.Matches;
+import es.codeurjc.backend.model.Player;
 import es.codeurjc.backend.model.Team;
+import es.codeurjc.backend.service.MatchService;
+import es.codeurjc.backend.service.PlayerService;
 import es.codeurjc.backend.service.TeamService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,6 +25,10 @@ import java.util.Optional;
 public class TeamRestController {
     @Autowired
     private TeamService teamService;
+    @Autowired
+    private PlayerService playerService;
+    @Autowired
+    private MatchService matchService;
 
     @GetMapping
     public ResponseEntity<List<Team>> getAllTeams(@RequestParam(defaultValue = "0") int page,
@@ -58,8 +68,26 @@ public class TeamRestController {
         return ResponseEntity.ok(updatedTeamDTO);
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTeam(@PathVariable Long id){
-        teamService.deleteTeamById(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<?> deleteTeam(@PathVariable Long id) {
+        try {
+            List<Player> players = playerService.findPlayersTeamById(id);
+
+            for (Player player : players) {
+                player.setTeam(null);
+                playerService.save(player);
+            }
+
+            List<Matches> matches = matchService.findMatchesByTeamId(id);
+            if (!matches.isEmpty()) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                        .body("No se puede eliminar el equipo porque tiene partidos asociados.");
+            }
+
+            teamService.deleteTeamById(id);
+            return ResponseEntity.noContent().build();
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
+
 }
